The operation I named AuRaMan (Autonomous Radiation Measurement) maps radiation of a confined surface using a Geiger Muller Counter mounted on an autonomous vehicle .
The vehicle starts from a corner of a room and moves in a straight row ,collecting radiation count every 1 sec .These data are added as entries of a row vector. When a wall is sensed at front, it takes another row parallel to the previous one. Several row vectors are created in the same manner. The row vectors are transmitted to a computing device through a Bluetooth module(ZS-040). All the row vectors may not have the same number of entries. Thus sophisticated algorithms were used to make the entry numbers equal to the entry number of the first row vector  . Therefore,, these row vectors are added up to a rectangular matrix  using numpy and plotted as radiation map using matplotlib. This project is yet to be deployed ,thus I used random numbers to simulate radiation data and mechanical complications (for example variable number of entries per row)